docs: rename files auth usecase
